Added better MaxPrimitive value for OpenGL using the gfx cards internal value. Changed some return values from s32 to u32 to avoid check for <0. Added 24bit depth window test for OpenGL under Windows.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@433 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CMY3DMeshFileLoader.cpp
@@ -602,7 +602,7 @@ for (u32 p=0; p<nToRead; p++)
 		SMeshBufferLightMap* buffer = getMeshBufferByMaterialIndex(meshHeader.MatIndex);
 
 		if (!buffer ||
-			((int)buffer->Vertices.size()+vertsNum) > Driver->getMaximalPrimitiveCount())
+			(buffer->Vertices.size()+vertsNum) > Driver->getMaximalPrimitiveCount())
 		{
 			// creating new mesh buffer for this material
 			buffer = new scene::SMeshBufferLightMap();
